requirements.txt
python-telegram-bot
httpx
parsel
requests
tqdm


